% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cirWrappers.r
\name{drplot}
\alias{drplot}
\title{Visualizing the dose-response summary of an up-and-down experiment}
\usage{
drplot(
  x,
  y,
  shape = "X",
  connect = FALSE,
  symbcol = 1,
  addest = FALSE,
  addcurve = FALSE,
  target = NULL,
  balancePt = target,
  estcol = "purple",
  estsize = 2,
  estsymb = 19,
  esthick = 2,
  curvecol = "blue",
  ytitle = "Frequency of Positive Response",
  xtitle = "Dose / Stimulus",
  ...
)
}
\arguments{
\item{x}{numeric vector: sequence of administered doses, treatments, stimuli, etc.}

\item{y}{numeric vector: sequence of observed responses. Must be same length as \code{x}, and must be coded \code{TRUE/FALSE} or 0/1.}

\item{shape}{the plotting shape (DRtrace only): \code{'circle'} (default), \code{'square'}, or \code{'triangle'}.}

\item{connect}{logical: whether to connect the symbols (generic plotting type \code{'b'}). Default \code{TRUE} for \code{udplot()} and \code{FALSE} for \code{drplot()}.}

\item{symbcol}{The color of the main plotting symbols and connecting lines. Default 1 (the current palette's first color). Note: if you change the color and inadvertently use \code{col} instead, there might be an error message.}

\item{target}{The target response rate for which target dose estimate is requested. Must be a single number in \eqn{(0,1).}}

\item{balancePt}{In case the design's inherent balance point differs somewhat from \code{target}, specify it here to improve estimation accuracy. See Details for further explanation. Otherwise, this argument defaults to be equal to \code{target}.}

\item{xtitle, ytitle}{x-axis and y-axis titles. Some reasonable defaults are provided, to avoid an annoying error message.}

\item{...}{Other arguments passed on to \code{\link[graphics]{plot}} (e.g., \code{main} for the main title).}
}
\description{
Visualizing the dose-response summary of an up-and-down experiment
}
\seealso{
\itemize{
\item \code{\link[cir]{plot.doseResponse}}, \code{cir} package.
\item \code{cir} package vignette.
}
}
\author{
Assaf P. Oron \code{<assaf.oron.at.gmail.com>}
}
