% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{krow}
\alias{krow}
\title{Up-and-Down Design Rules for use in Dose-Finding Simulator}
\usage{
krow(doses, responses, k, lowTarget = NULL, cohort = 1, fastStart = FALSE, ...)
}
\arguments{
\item{doses, responses}{(mandatory arguments) vectors of the run's current sequence of doses (in ordinal/index scale) and responses}

\item{lowTarget}{(\code{krow} and \code{bcd}) logical: is the target below 0.5 (median threshold)?}

\item{fastStart}{(\code{krow} and \code{bcd}) logical: should the experiment begin with a classical-UD-like stage until the first "minority" response is observed (i.e., a 1 for below-median targets and vice versa)? Even though \code{TRUE} delivers better experimental performance and is recommended when allowed, default is \code{FALSE} because toxicity/safety studies are unlikely to allow it.}
}
\description{
Rules for k-in-a-row, Biased-Coin UD, and Group UD, coded as functions compatible with
the generic dose-finding simulator \code{dfsim()}

These functions work on each virtual experimental run individually.
}
\details{
Rules for some popular or well-studied non-up-and-down

@return the next dose allocation
}
