% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorFunctions.r
\name{pivec}
\alias{pivec}
\alias{currentvec}
\alias{cumulvec}
\title{Key Probability Vectors of Up-and-Down Designs}
\usage{
pivec(cdf, matfun, ...)

currentvec(startdose = NULL, cdf, n, matfun, ...)

cumulvec(startdose = NULL, cdf, n, matfun, average = TRUE, exclude = 0, ...)
}
\arguments{
\item{cdf}{vector of positive-response probabilities at the doses}

\item{matfun}{The function to calculate the TPM. Depends on the specific design; see \code{\link{bcdmat}}. Must}
}
\description{
Asymptotic and progression
}
\details{
Up-and-Down designs (UDDs) generate random walk behavior, whose theoretical properties can be summarized via a transition probability matrix (TPM). Given the number of doses \eqn{M}, and the value of the cdf \eqn{F} at each dose (i.e., the positive-response probabilities), the specific UDD rules uniquely determine the TPM.
}
\note{
When using the k-in-a-row design, set \code{matfun = kmatMarg}, not \code{kmatFull}.
}
\author{
Assaf P. Oron \code{<assaf.oron.at.gmail.com>}
}
