% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averaging.r
\name{reversals}
\alias{reversals}
\title{Identify reversals in an UD experiment's time series}
\usage{
reversals(y)
}
\arguments{
\item{y}{time series of binary responses, should be coded 0/1 or \code{FALSE/TRUE}}
}
\description{
Identify reversals in an UD experiment's time series
}
