% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averaging.r
\name{reversmean}
\alias{reversmean}
\title{Reversal-anchored averaging estimators: reversal-only and reversal-startpoint}
\usage{
reversmean(
  x,
  y,
  rstart = 3,
  all = TRUE,
  minfrac = 0.5,
  before = 0,
  full = FALSE
)
}
\description{
Reversal-anchored averaging estimators: reversal-only and reversal-startpoint
}
