% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.r
\name{udboot}
\alias{udboot}
\title{Generic percentile dose-response / dose-finding bootstrap routine}
\usage{
udboot(
  x,
  y,
  doses = NULL,
  estfun = dynamean,
  design = krow,
  desArgs = list(k = 1),
  target = 0.5,
  conf = 0.9,
  B = 1000,
  seed = NULL,
  randstart = TRUE,
  showdots = TRUE,
  full = FALSE,
  ...
)
}
\arguments{
\item{x}{numeric vector: sequence of administered doses, treatments, stimuli, etc.}

\item{y}{numeric vector: sequence of observed responses. Must be same length as \code{x} or shorter by 1, and must be coded \code{TRUE/FALSE} or 0/1. \code{dynamean()} only uses \code{y} for bootstrap confidence intervals.}

\item{doses}{the complete set of dose values that \emph{could} have been included in the experiment. Must include all unique values in \code{x}.}
}
\description{
Bootstrap routine for resampling a dose-finding or dose-response experiment. The bootstrap replicates are generated from a centered-isotonic-regression estimate of the dose-response function, rather than resampled directly.
}
\seealso{
\code{\link{simulate}}
}
\author{
Assaf P. Oron \code{<assaf.oron.at.gmail.com>}
}
