% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cirWrappers.r
\name{udest}
\alias{udest}
\title{Centered-Isotonic-Regression (CIR) Estimate for the Up-and-Down Target Dose}
\usage{
udest(x, y, target, balancePt = target, conf = 0.9)
}
\arguments{
\item{x}{numeric vector: sequence of administered doses, treatments, stimuli, etc.}

\item{y}{numeric vector: sequence of observed responses. Must be same length as \code{x} or shorter by 1, and must be coded \code{TRUE/FALSE} or 0/1.}

\item{target}{The target response rate for which target dose estimate is requested. Must be a single number in $(0,1).$}

\item{balancePt}{In case the design's inherent balance point differs somewhat from \code{target}, specify it here to improve estimation accuracy. See Details for further explanation. Otherwise, this argument defaults to be equal to \code{target}}

\item{conf}{The desired confidence level for the confidence interval. Default $90\\%,$ and we recommend not increasing to $95\\%$ unless you have about $\sim 100$ or more observations.
https://cdn-links.lww.com/permalink/aln/c/aln_2022_05_25_oron_aln-d-21-01101_sdc1.pdf}
}
\description{
CIR is the recommended method for estimating up-and-down (UD) targets. This is a UD-adapted convenience wrapper for a function from the \code{cir} package.
}
\details{
Centered Isotonic Regression (CIR) is an extension of isotonic regression (IR), substantially improving upon IR's estimation performance in the dose-response and dose-finding contexts (Oron and Flournoy 2017, Flournoy and Oron 2020). CIR and related methods are available in the \code{cir} package. The \code{udest()} function in the present package provides a convenient wrapper for \code{cir::quickInverse()}, with arguments already set to the appropriate values for estimating the target dose after an up-and-down experiment.

WARNING! You should not estimate
}
\author{
Assaf P. Oron \code{<assaf.oron.at.gmail.com>}
}
