% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cirWrappers.r
\name{udplot}
\alias{udplot}
\title{Visualizing the time series of an up-and-down experiment}
\usage{
udplot(
  x,
  y,
  shape = "circle",
  connect = TRUE,
  symbcol = 1,
  doselabels = NULL,
  xtitle = "Observation Order",
  ytitle = "Dose / Stimulus",
  ...
)
}
\arguments{
\item{x}{numeric vector: sequence of administered doses, treatments, stimuli, etc.}

\item{y}{numeric vector: sequence of observed responses. Must be same length as \code{x}, and must be coded \code{TRUE/FALSE} or 0/1.}

\item{shape}{the plotting shape (DRtrace only): \code{'circle'} (default), \code{'square'}, or \code{'triangle'}.}

\item{connect}{logical: whether to connect the symbols (generic plotting type \code{'b'}). Default \code{TRUE} for \code{udplot()} and \code{FALSE} for \code{drplot()}.}

\item{symbcol}{The color of the main plotting symbols and connecting lines. Default 1 (the current palette's first color). Note: if you change the color and inadvertently use \code{col} instead, there might be an error message.}

\item{doselabels}{(\code{DRtrace} only) Dose values to be plotted along the y-axis. If \code{NULL} (default), those will be the doses in the dataset (i.e., \code{sort(unique(x))}).}

\item{xtitle, ytitle}{x-axis and y-axis titles. Some reasonable defaults are provided, to avoid an annoying error message.}

\item{...}{Other arguments passed on to \code{\link[graphics]{plot}} (e.g., \code{main} for the main title).}
}
\description{
Plotting function for the "trace" (time series) of an up-and-down experiment, showing the observation order on the x-axis, and the dose \emph{(treatment, stimulus, etc.)} strength on the y-axis. Uses utilities from the \code{cir} package.
}
\details{
This simple and handy visualization approach was presented already by Dixon and Mood (1948).
\itemize{
\item It conveys directly the meaning of \emph{"up-and-down"}, because the administered dose/stimulus strength is on the y-axis, whereas observation order is on the x-axis.
\item Filled symbols stand for positive responses and open symbols for negative.
\item The design's transition rules can be usually inferred directly from the plot.
}

\code{udplot()} is a convenience wrapper to \code{cir::plot.DRtrace}. This is a base-R plot, so you can use additional options, including preceding the plot command with \code{\link[graphics]{par}} statements, or following up with \code{\link[graphics]{legend}}. When wishing to save to a file, I recommend utilities such as \code{png()} or \code{pdf()}.
}
\examples{

## This example is from one of the earliest applications of up-and-down in anesthesiology:
## Columb MO, Lyons, G. 1995. Determination of the Minimum Local Analgesic Concentrations of Epidural Bupivacaine and Lidocaine in Labor, Anesthesia & Analgesia 81, 833-837.
## open access: https://journals.lww.com/anesthesia-analgesia/Fulltext/1995/10000/Determination_of_the_Minimum_Local_Analgesic.30.aspx

# It was a classical median-finding up-and-down study.

# Bupivacaine: 
bupix = 0.01 * c(15:9,10:6,7,6,7:5,6,7:4,5,6,5:8,7,6)
# With classical U&D, responses (except the last one) can be read off the doses:
bupiy = c( (1 - sign(diff(bupix)))/2, 0 )

# Lidocaine (note that spacing was halved after observation 5):
lidox = 0.05 * c(seq(20,12,-2),11:9,10:7,8,7,8,7:9,8:6,7,8,7,8:5,6,7)
lidoy = c( (1 - sign(diff(lidox)))/2, 0 )

### Plots plots plots!

layout(matrix(1:4, nrow=2, byrow=TRUE), widths=3:2)
par(mar=c(4,4,4,1), mgp=c(2.5,0.8,0), cex.axis = 0.7, las = 1)

# The experimental trajectory / time-series / "trace" (pick your favorite name!)
# Note the changed argument names for x and y axis titles
udplot(bupix, bupiy, main='Columb and Lyons 95: Bupivacaine', 
        xtitle = "Patient Number", ytitle = 'Bupivacaine (\% w/v)') 
# Compare with the article's Figure 1; the commented line below makes it look more similar
# udplot(bupix, bupiy, shape='square', connect=FALSE, cex=2)

# The dose-response plot, rarely encountered in U&D articles. 
#   (Columb and Lyons do provide the info as a table, sort-of - see Table 2)
# We can also add the CIR estimate right there:
drplot(bupix, bupiy, main='Bupivacaine Dose-Response', percents = TRUE,
       addest = TRUE, target = 0.5, addcurve = TRUE,
       xtitle = 'Bupivacaine (\% w/v)', ytitle = "Percent Effective")

# Same two plots for lidocaine

udplot(lidox, lidoy, main='Columb and Lyons 95: Lidocaine',
  xtitle = "Patient Number", ytitle = 'Lidocaine (\% w/v)') 

drplot(lidox, lidoy, main='Lidocaine Dose-Response', percents = TRUE,
       addest = TRUE, target = 0.5, addcurve = TRUE,
       xtitle = 'Bupivacaine (\% w/v)', ytitle = "Percent Effective")

#----------- Let us actually see the numbers of those Centered-Isotonic-Regression (CIR) estimates!
# Note that our default confidence-interval is 90\%. Change it via the 'conf' argument.

udest(bupix, bupiy, target = 0.5)
# Compare with the article: 0.065\% (95\% CI 0.045-0.085) via "Dixon-Massey" (presumably, Dixon-Mood), 
#                       and 0.062\% (95\% CI 0.052-0.072) via logistic regression

udest(lidox, lidoy, target = 0.5)
# Compare with the article: 0.37\% (95\% CI 0.32-0.42) via "Dixon-Massey"  
#                       and 0.36\% (95\% CI 0.31-0.41) via logistic regression

# Our package does include the Dixon-Mood point estimate.
#  (w/o the CIs, because we do not endorse this estimation approach)
# Does it reproduce the article numbers?
dixonmood(bupix, bupiy)
dixonmood(lidox, lidoy)
# Not quite. Perhaps indeed the Dixon-Massey textbook (published a few years after Dixon and Mood)
#    included some tweak on the original. Or perhaps, the formula is a bit open to variations in
#    translation when coding it (originally it was a pencil-and-paper affair).


}
\seealso{
\itemize{
\item \code{\link[cir]{plot.DRtrace}}, \code{cir} package.
\item \code{\link{drplot}} for the up-and-down dose-response and estimate plotting.
\item \code{cir} package vignette.
}

#' @references
\itemize{
\item Dixon WJ, Mood AM. A method for obtaining and analyzing sensitivity data. \emph{J Am Stat Assoc.} 1948;43:109-126.
\item Oron AP, Souter MJ, Flournoy N. Understanding Research Methods: Up-and-down Designs for Dose-finding. \emph{Anesthesiology} 2022; 137:137–50.
}
}
\author{
Assaf P. Oron \code{<assaf.oron.at.gmail.com>}
}
