% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averaging.r
\name{adaptmean}
\alias{adaptmean}
\title{Average with adaptive rstarting-point
Based on an unpblished concept from Oron (2007)}
\usage{
adaptmean(x, minfrac = 2/3, before = FALSE, full = FALSE)
}
\description{
Average with adaptive rstarting-point
Based on an unpblished concept from Oron (2007)
}
