% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixFunctions.r
\name{bcdmat}
\alias{bcdmat}
\title{Transition Probability Matrix Utilities}
\usage{
bcdmat(cdf, target)
}
\arguments{
\item{cdf}{monotone increasing vector with positive-response probabilities. The number of dose levels $M$ is deduced from vector's length.}

\item{target}{the design's target response rate.}

\item{repeatNegatives}{logical, relevant only for k-in-a-row designs: are repeated negative responses needed to level up, or vice versa (repeated positives to level down)? See "Details" for more information.}
}
\value{
for `bcdmat` and `kmatMarg`, an $M\times M$ transition probability matrix. For `pivec`,  an $M$-length vector with stationary/asymptotic visit frequencies.
}
\description{
Transition Probability Matrix Utilities
}
