% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averaging.r
\name{dixonmood}
\alias{dixonmood}
\title{Original Dixon and Mood (1948) point estimate}
\usage{
dixonmood(x, y, full = FALSE, flip = FALSE)
}
\arguments{
\item{x}{numeric vector: sequence of administered doses, treatments, stimuli, etc.}

\item{y}{numeric vector: sequence of observed responses. Must be same length as \code{x} or shorter by 1, and must be coded \code{TRUE/FALSE} or 0/1.}

\item{full}{logical: should more detailed information be returned, or only the estimate? (default \code{FALSE})}

\item{flip}{logical: should we flip D-M's approach and use the more-common outcome? (default \code{FALSE})}
}
\description{
Basic version; formula assumes uniform spacing but should work anyway
}
\details{
In their documentation of the Up-and-Down algorithm, Dixon and Mood (1948) presented an estimation method based on tallying responses
}
