% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cirWrappers.r
\name{udest}
\alias{udest}
\title{Centered-Isotonic-Regression (CIR) Estimate for the Up-and-Down Target Dose}
\usage{
udest(x, y, target, balancePt = target, conf = 0.9, ...)
}
\arguments{
\item{x}{numeric vector: sequence of administered doses, treatments, stimuli, etc.}

\item{y}{numeric vector: sequence of observed responses. Must be same length as \code{x}, and must be coded \code{TRUE/FALSE} or 0/1.}

\item{target}{The target response rate for which target dose estimate is requested. Must be a single number in \eqn{(0,1).}}

\item{balancePt}{In case the design's inherent balance point differs somewhat from \code{target}, specify it here to improve estimation accuracy. See Details for further explanation. Otherwise, this argument defaults to be equal to \code{target}.}

\item{conf}{The desired confidence level for the confidence interval. Default \eqn{90\%.} We do not recommend increasing to \eqn{95\%} unless you have \eqn{\sim 100} or more observations.}

\item{...}{Pass-through argument added for flexible calling context.}
}
\value{
A one-row data frame with 4 variables: \code{target}, \code{point} (the point estimate), \verb{lowerXYconf, upperXYconf} (the confidence bounds, with \code{XY} standing for the percents, default \code{90}).
}
\description{
Centered Isotonic Regression (CIR) is an extension of isotonic regression (IR), substantially improving upon IR's estimation performance in the dose-response and dose-finding contexts (Oron and Flournoy 2017, Flournoy and Oron 2020). CIR is the recommended method for estimating up-and-down targets.
}
\details{
CIR and related methods are available in the \code{cir} package. The \code{udest()} function in the present package provides a convenient wrapper for \code{cir::quickInverse()}, with arguments already set to the appropriate values for estimating the target dose after an up-and-down experiment. The function also returns a confidence interval as default.

\strong{WARNING!} You should not estimate target doses too far removed from the design's actual balance point (definitely no further than 0.1, e.g., estimating the 33rd percentile for a design whose balance point is the median). As Flournoy and Oron (2020) explain, observed response rates are biased away from the balance point. Even though \code{udest()} performs the rudimentary bias correction described in that article, practically speaking this correction's role is mostly to expand the confidence intervals in response to the bias. It cannot guarantee to provide reliable off-balance-point estimates.
}
\examples{

## This example is from one of the earliest applications of up-and-down in anesthesiology:
## Columb MO, Lyons, G. 1995. Determination of the Minimum Local Analgesic Concentrations of Epidural Bupivacaine and Lidocaine in Labor, Anesthesia & Analgesia 81, 833-837.
## open access: https://journals.lww.com/anesthesia-analgesia/Fulltext/1995/10000/Determination_of_the_Minimum_Local_Analgesic.30.aspx

# This was a classical median-finding up-and-down study.

# Bupivacaine: 
bupix = 0.01 * c(15:9,10:6,7,6,7:5,6,7:4,5,6,5:8,7,6)
# With classical U&D, responses (except the last one) can be read off the doses:
bupiy = c( (1 - sign(diff(bupix)))/2, 0 )

# Lidocaine (note that spacing was halved after observation 5):
lidox = 0.05 * c(seq(20,12,-2),11:9,10:7,8,7,8,7:9,8:6,7,8,7,8:5,6,7)
lidoy = c( (1 - sign(diff(lidox)))/2, 0 )

### Plots plots plots!

layout(matrix(1:4, nrow=2, byrow=TRUE), widths=3:2)
par(mar=c(4,4,4,1), mgp=c(2.5,0.8,0), cex.axis = 0.7, las = 1)

# The experimental trajectory / time-series / "trace" (pick your favorite name!)
# Note the changed argument names for x and y axis titles
udplot(bupix, bupiy, main='Columb and Lyons 95: Bupivacaine', 
        xtitle = "Patient Number", ytitle = 'Bupivacaine (\% w/v)') 
# Compare with the article's Figure 1; the commented line below makes it look more similar
# udplot(bupix, bupiy, shape='square', connect=FALSE, cex=2)

# The dose-response plot, rarely encountered in U&D articles. 
#   (Columb and Lyons do provide the info as a table, sort-of - see Table 2)
# We can also add the CIR estimate right there:
drplot(bupix, bupiy, main='Bupivacaine Dose-Response', percents = TRUE,
       addest = TRUE, target = 0.5, addcurve = TRUE,
       xtitle = 'Bupivacaine (\% w/v)', ytitle = "Percent Effective")

# Same two plots for lidocaine

udplot(lidox, lidoy, main='Columb and Lyons 95: lidocaine',
  xtitle = "Patient Number", ytitle = 'Lidocaine (\% w/v)') 

drplot(lidox, lidoy, main='Lidocaine Dose-Response', percents = TRUE,
       addest = TRUE, target = 0.5, addcurve = TRUE,
       xtitle = 'Bupivacaine (\% w/v)', ytitle = "Percent Effective")

#----------- Let us actually see the numbers of those Centered-Isotonic-Regression (CIR) estimates!
# Note that our default confidence-interval is 90\%. Change it via the 'conf' argument.

udest(bupix, bupiy, target = 0.5)
# Compare with the article: 0.065\% (95\% CI 0.045-0.085) via "Dixon-Massey" (presumably, Dixon-Mood), 
#                       and 0.062\% (95\% CI 0.052-0.072) via logistic regression

udest(lidox, lidoy, target = 0.5)
# Compare with the article: 0.37\% (95\% CI 0.32-0.42) via "Dixon-Massey" (presumably, Dixon-Mood), 
#                       and 0.36\% (95\% CI 0.31-0.41) via logistic regression

# Our package does include the Dixon-Mood point estimate 
#  (w/o the confidence, because we do not endorse this estimation approach)
# Does it reproduce the article numbers?
dixonmood(bupix, bupiy)
dixonmood(lidox, lidoy)
# Not quite. Perhaps indeed the Dixon-Massey textbook (published a few years after Dixon and Mood)
#    included some tweak on the original. Or perhaps, the formula is a bit open to variations in
#    translation when coding it (originally it was a pencil-and-paper affair).
}
\references{
\itemize{
\item Oron AP, Flournoy N.  Centered Isotonic Regression: Point and Interval Estimation for Dose-Response Studies. \emph{Statistics in Biopharmaceutical Research} 2017; 9, 258-267. \href{https://arxiv.org/pdf/1701.05964.pdf}{Author's public version available on arxiv.org.}
\item Flournoy N, Oron AP. Bias Induced by Adaptive Dose-Finding Designs. \emph{Journal of Applied Statistics} 2020; 47, 2431-2442.
\item Oron AP, Souter MJ, Flournoy N. Understanding Research Methods: Up-and-down Designs for Dose-finding. \emph{Anesthesiology} 2022; 137:137–50. \href{https://cdn-links.lww.com/permalink/aln/c/aln_2022_05_25_oron_aln-d-21-01101_sdc1.pdf}{See in particular the open-access Supplement.}
}

@seealso
\itemize{
\item \code{\link[cir]{quickInverse}}, \code{cir} package.
\item \code{cir} package vignette.
}
}
\author{
Assaf P. Oron \code{<assaf.oron.at.gmail.com>}
}
