% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cirWrappers.r
\name{udplot}
\alias{udplot}
\title{Visualizing the time series of an up-and-down experiment}
\usage{
udplot(
  x,
  y = NULL,
  shape = "circle",
  connect = TRUE,
  symbcol = 1,
  doselabels = NULL,
  xtitle,
  ytitle,
  ...
)
}
\arguments{
\item{x}{numeric vector: sequence of administered doses, treatments, stimuli, etc.}

\item{y}{numeric vector: sequence of observed responses. Must be same length as \code{x}, and must be coded \code{TRUE/FALSE} or 0/1.}

\item{shape}{the plotting shape (DRtrace only): 'circle' (default), 'square', or 'triangle'}

\item{connect}{logical: whether to connect the symbols (generic plotting type 'b'). Default \code{TRUE} for \code{\link{DRtrace}} and \code{FALSE} for \code{\link{doseResponse}}}

\item{symbcol}{The color of the main plotting symbols and connecting lines. Default 1 (the current palette's first color). Note: if you change the color and inadvertently use \code{col} instead, there might be an error message.}

\item{doselabels}{(\code{DRtrace} only) Dose values to be plotted along the y-axis. If \code{NULL} (default), those will be the doses in the dataset (i.e., \code{sort(unique(x))}).}

\item{xtitle, ytitle}{x-axis and y-axis titles}

\item{...}{Other arguments passed on to \code{\link{plot}}.}
}
\description{
Visualizing the time series of an up-and-down experiment
}
\author{
Assaf P. Oron \code{<assaf.oron.at.gmail.com>}
}
